+++ src/main/java/com/example/luna/entity/OrderItem.java
@@ -19,5 +19,5 @@ public class OrderItem {
 
     private Integer quantity;
 
-    private Integer price;
+    private Float price;
 }
